# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management System API",
        "description": "Hello my name is Tresten Pool aspiring backend developer. I created this api for my advanced software engineering class",
        "contact": {
            "email": "trestenpool@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ec2-54-90-227-153.compute-1.amazonaws.com/api/"
        }
    ],
    "paths": {
...
        "/manufacturer": {
            "get": {
                "tags": [
                    "Manufacturers"
                ],
                "summary": "Get all the manufacturers",
                "operationId": "getManufacturerList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Manufacturers"
                ],
                "summary": "Store a new manufacturer",
                "operationId": "storeNewManufacturer",
                "requestBody": {
                    "description": "Pass in the manufacturer name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "manufacturer": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "default": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "409": {
                        "description": "Conflict, there is already a manufacturer with that name"
                    }
                }
            }
        },
        "/manufacturer/{manufacturer_id}": {
            "get": {
                "tags": [
                    "Manufacturers"
                ],
                "summary": "Get manufacturer by manufacturer_id",
                "operationId": "GetManufacturer",
                "parameters": [
                    {
                        "name": "manufacturer_id",
                        "in": "path",
                        "description": "Manufacturer id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "The manufacturer id route parameter is not valid. Must be numeric.."
                    },
                    "404": {
                        "description": "manufacturer not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Manufacturers"
                ],
                "summary": "Update manufacturer by manufacturer_id",
                "operationId": "updateManufacturer",
                "parameters": [
                    {
                        "name": "manufacturer_id",
                        "in": "path",
                        "description": "Manufacturer id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass in the manufacturer name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "manufacturer": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "default": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "404": {
                        "description": "manufacturer not found"
                    },
                    "409": {
                        "description": "Conflict, there is already a manufacturer with that name"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Manufacturers"
                ],
                "summary": "Delete manufacturer by manufacturer_id",
                "operationId": "DeleteManufacturer",
                "parameters": [
                    {
                        "name": "manufacturer_id",
                        "in": "path",
                        "description": "Manufacturer id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "manufacturer not found with the given ID"
                    }
                }
            }
        },
...
    }
}